+++ src/extract.rs
@@ -4,4 +4,5 @@ pub use crate::body_hash::BodyHash;
 pub use crate::body_hmac::{BodyHmac, HmacConfig};
 pub use crate::json::{Json, DEFAULT_JSON_LIMIT};
 pub use crate::lazy_data::LazyData;
+pub use crate::path::Path;
 pub use crate::query::Query;